Set max body size
As reported in TG-566, restd would stop working
after executing a POST, PUT or PATCH request with
a very big string as an attribute. This fix adds a
maximum body size to prevent this and a BAD REQUEST
status code is issued if the limit is exceeded.

TG-566
Tags: fix, usr

Change-Id: Id26f7f915181b3d17bb6d92b9e963b0c62067dfe
Signed-off-by: Ederick Colmenares <ederick.colmenares.sanchez@hpe.com>