Reduce IDL subscription to config columns for REST and some fixes.
IDL for REST is reduced to subscribe to configuration columns only.
Status/statistics are fetched on-demand to reduce the notification
overhead. Also fixes a REST transaction issue that may cause
transactions to hang when there is no IDL activity. Timeout
mechanism added for REST transaction processing.

REST get.py module modified to be executed in a coroutine since
the on-demand fetch logic utilizes a commit_block.

Change-Id: I14d0f79c2b648c30049c6e598013439372a766e6
Signed-off-by: Chinh Le <chinh.le@hpe.com>