Make OVSReference a subclass of OVSColumn in restparser
OVSReferece and OVSColumn differ in a small amount of attributes and
there were even some references that required all the same information
from a normal column.
This refactor makes OVSReference a sub class of OVSColumn, improving
future maintainability and making easier to support the new modular
schema format that will be introduced in the near future.

Tags: chg, dev

Change-Id: I486b804251da9628d5baa25260a93e33a545eec6
Signed-off-by: Eduardo Ramirez <eramireza@hpe.com>